[Design] 장바구니 라벨 추가
- UILabel을 추가하여 장바구니 표시
- 제약설정 추가
